Transfer List UI

UI Shiny Transfer List

mui_transferList(
  inputId,
  value = NULL,
  left_list = c("A", "B", "C"),
  right_list = c("D", "E", "F"),
  left_title = "Choices",
  right_title = "Chosen",
  direction = "row",
  justifyContent = "center"
)

Arguments

inputId
Input id

value
Default value (will be the list on the right by default)

left_list
Values of the list on the left (vector or list) (must be initialized with at least two values)

right_list
Values of the list on the right (vector or list) (must be initialized with at least two values)

left_title
The title of the list on the left (default = "Choices")

right_title
The title of the list on the right (default = "Chosen")

direction
The layout of the lists c("row", "column"). Default = "row"

justifyContent
justifyContent prop passed to the main grid. options: flex-start, center, flex-end, space-between, space-around, space-evenly

See also

Examples

mui_transferList("transferList_1", left_list = names(mtcars),
right_list = names(iris))
#> <span id="transferList_1" class="mui_transferlist"></span>
#> <script id="transferList_1_value" type="application/json">["Sepal.Length","Sepal.Width","Petal.Length","Petal.Width","Species"]</script>
#> <script id="transferList_1_configuration" type="application/json">{"left_list":["mpg","cyl","disp","hp","drat","wt","qsec","vs","am","gear","carb"],"right_list":["Sepal.Length","Sepal.Width","Petal.Length","Petal.Width","Species"],"left_title":"Choices","right_title":"Chosen","direction":"row","justifyContent":"center"}</script>